pyproject.toml
[project]
name = "financial-researcher-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=1.6.1",
    "crewai-tools>=1.6.1",
    "litellm>=1.80.7",
    "numpy>=2.3.5",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.52.0",
]
